+++ cardano-node/src/Cardano/Node/Tracing/Tracers/Consensus.hs
@@ -72,6 +72,10 @@ module Cardano.Node.Tracing.Tracers.Consensus
   , severityKeepAliveClient
   , docKeepAliveClient
 
+  , namesConsensusStartupError
+  , severityConsensusStartupError
+  , docConsensusStartupError
+
   ) where
 
 
@@ -93,8 +97,9 @@ import           Cardano.Node.Tracing.Era.Byron ()
 import           Cardano.Node.Tracing.Era.Shelley ()
 import           Cardano.Node.Tracing.Formatting ()
 import           Cardano.Node.Tracing.Render
+import           Cardano.Node.Tracing.Tracers.ConsensusStartupException
+import           Cardano.Node.Tracing.Tracers.ForgingThreadStats (ForgingStats)
 import           Cardano.Node.Tracing.Tracers.StartLeadershipCheck
-import           Cardano.Node.Tracing.Tracers.ForgingThreadStats(ForgingStats)
 import           Cardano.Prelude hiding (All, Show, show)
 
 import           Cardano.Protocol.TPraos.OCert (KESPeriod (..))
@@ -130,7 +135,6 @@ import           Ouroboros.Consensus.Node.Run (SerialiseNodeToNodeConstraints, e
 import           Ouroboros.Consensus.Node.Tracers
 import qualified Ouroboros.Consensus.Protocol.Ledger.HotKey as HotKey
 import           Ouroboros.Consensus.Util.Enclose
-
 
 
 instance LogFormatting a => LogFormatting (TraceLabelCreds a) where
@@ -577,7 +581,7 @@ instance (HasHeader header, ConvertRawHash header) =>
     mconcat [ "kind" .= String "AddedFetchRequest" ]
   forMachine _dtal BlockFetch.AcknowledgedFetchRequest {} =
     mconcat [ "kind" .= String "AcknowledgedFetchRequest" ]
-  forMachine _dtal (BlockFetch.SendFetchRequest af) =
+  forMachine _dtal (BlockFetch.SendFetchRequest af _) =
     mconcat [ "kind" .= String "SendFetchRequest"
             , "head" .= String (renderChainHash
                                  (renderHeaderHash (Proxy @header))
@@ -1711,6 +1715,20 @@ instance Show remotePeer => LogFormatting (TraceKeepAliveClient remotePeer) wher
 
 docKeepAliveClient :: Documented (TraceKeepAliveClient peer)
 docKeepAliveClient = Documented [
+    DocMsg
+      []
+      []
+      ""
+  ]
+
+namesConsensusStartupError :: ConsensusStartupException -> [Text]
+namesConsensusStartupError _ = []
+
+severityConsensusStartupError :: ConsensusStartupException -> SeverityS
+severityConsensusStartupError _ = Critical
+
+docConsensusStartupError :: Documented ConsensusStartupException
+docConsensusStartupError = Documented [
     DocMsg
       []
       []