+++ cardano-node/ChangeLog.md
@@ -1,6 +1,26 @@
 # Changelog for cardano-node
 
-## vNext
+## 1.36.0 -- November 2022
+
+### node changes
+
+ - Bumped cardano-ledger to 0.1
+ - Bumped plutus-core to 1.0
+ - Bumped ouroboros-network to commit 72863b0
+
+### consensus changes
+
+TODO
+
+### network changes
+
+TODO
+
+### ledger changes
+
+TODO
+
+## 1.35.5 -- November 2022
 
 ### node changes
 
@@ -12,13 +32,41 @@ None
 
 ### network changes
 
-- New P2P topology file format. See [issue #4563][#4563] or the [config
-  files][understanding-config-files] documentation. The old P2P topology
+- Added 'DemoteLocalAsynchronous' warning trace.  It indicates that a remote
+  local root peer was demoted to cold (either due to connection error or
+  misbehaviour).
+
+- New P2P topology file format, see [issue #4563][#4563] or the [config
+  files][understanding-config-files] documentation.  The old p2p topology
   format will be supported for next two major releases of the node (the last
-  supported major version will be `1.37`).
+  major version which will support it is `1.37`).
 
 [#4563]: https://github.com/input-output-hk/cardano-node/issues/4563
 [understanding-config-files]: https://github.com/input-output-hk/cardano-node/blob/master/doc/getting-started/understanding-config-files.md
+
+### ledger changes
+
+None
+
+## 1.35.4 -- October 2022
+
+### node changes
+
+- Update plutus to version 1.0.0.1 to enable SECP at protocol version 8
+- Update cardano-crypto-class to version 2.0.0.0.1 to add SECP crypto primitives
+- Update block header advertised version in babbage to 8.0
+
+### plutus changes
+
+- enables SECP primitive for plutus smart contracts at protocol version 8 (intra-era hard fork)
+
+### consensus changes
+
+None
+
+### network changes
+
+None
 
 ### ledger changes